Modify behavior of the ClanStartup Patches to prevent possible undefined behaviour.
If FullClanMod is install there's a case to handle where MainClan == SubClan. this would grant StarterRelics, StarterCardUpgrade, and RandomDraftEnhancerPool twice. The first isn't a major concern as Wurmkin's starter relic does absolutely nothing, and applying Corruption to a card twice does nothing. The last is a behavior change which should be decided on a clan by clan basics for modders. The default behaviour should be apply effects once, even if the clans are the same since removing already added effects is harder than adding it again.